This works, but it has a few caveats.
We put the metal view INSIDE the content view, and update the metal view's
frame dynamically every time a scroll happens. This means that the metal
view stays stationary, from the perspective of the user, as we scroll.
Because we also update the metal view's orthographic content bounds, it
appears that the red boxes track the colored boxes.

We disable this entire process when isZoomBouncing = true, to get around
the snap back problem. This means that when the zoom is bouncing, the
metal view doesn't update except due to it's layer ancestry transform,
which means that the animation back to minimumZoomScale is smooth.

However, this also ends up with some issues:

1. Because the metal view gets very small when you zoom in enough, the
lines are blurry
2. Because the metal view is very small, the texture that metal is
rendering to is tiny, so pixel boudaries get hit. This means that when
zoomed in really far and you scroll slowly, the metal view only updates
the red frames when their "pixel" value changes, which is not that often
since there are no partial pixels.
3. When zoomed in all the way, the animation back to maximumZoomScale is
smooth, but the red frames get cut off because the metal view isn't
rendering them during zoom